Refactor the code, created components and hooks
